refactor : primitive type을 wrapper type으로 수정.

+++ src/main/java/com/scyllacore/dumpWeb/commonModule/db/dto/manage/DriverDTO.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class DriverDTO {
-    private int driverId;
+    private Long driverId;
     private String carNo;
     private String name;
-    private int userIdFk;
+    private Long userIdFk;
 }
